Redo the treatment of radial integrals
Previously, each integrand f(r) was pre-computed at the nodes (the
points of known (r_i, phi_i) from TFDH::solve). A spline was
constructed from the f_i to interpolate between these points and
obtain the general f(r) needed in the quadratures.

Now, a spline is constructed from the phi_i, giving arbitrary
phi(r). The code passes the integrand to the quadrature routine as
a lambda or functor, allowing for direct evaluation of f(r) at the
desired points.

This change was made to streamline the code, and facilitate possible
future changes to the interpolation (for example, interpolating in
log(r)-log(phi) space instead r-phi space).